requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
tinydb==4.8.2
tinydb-serialization==2.2.0
tinymongo==0.2.0
typing_extensions==4.14.1
yarl==1.20.1
